+++ de/Linus122/TelegramComponents/Chat.java
@@ -2,6 +2,6 @@ package de.Linus122.TelegramComponents;
 
 public class Chat {
 	public String text;
-	public int chat_id;
+	public long chat_id;
 	public String parse_mode;
 }
